불필요한 캐스팅 제거

+++ kdk/배열의평균값.java
@@ -13,6 +13,6 @@ public class 배열의평균값 {
 
     public double solution(int[] numbers) {
 
-        return (double)Arrays.stream(numbers).average().orElse(0);
+        return Arrays.stream(numbers).average().orElse(0);
     }
 }
